projects/core/gcp-oece-tech/data/legacy/repos/deepweay/apps/forum-backend/requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
google-generativeai==0.8.5
supabase==2.10.0
redis==5.2.1
aiohttp==3.11.11
pydantic==2.12.4
python-dateutil==2.9.0
